Create post method (Category)
