--- src/main/java/category/CategoryController.java
@@ -1,4 +1,0 @@
-package category;
-
-public class CategoryController {
-}
