--- src/main/java/category/CategoryRepository.java
@@ -1,9 +1,0 @@
-package category;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-
-    Category findOne(Integer id);
-    void delete(Integer id);
-}
