+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
 spring.jpa.hibernate.ddl-auto=create
 spring.datasource.url=jdbc:mysql://localhost:3306/kakeibo
 spring.datasource.username=kakeibodev
-spring.datasource.password=app1206dev
+spring.datasource.password=[password]
