+++ src/main/resources/templates/setting.html
@@ -3,9 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <title>KakeiboApp</title>
-    <link rel="stylesheet" href="../static/css/bootstrap.css" />
-    <link rel="stylesheet" href="../static/css/bootstrap.min.css" />
-    <link rel="stylesheet" href="../static/css/custom.css" />
+    <link rel="stylesheet" href="/css/bootstrap.css" />
+    <link rel="stylesheet" href="/css/bootstrap.min.css" />
+    <link rel="stylesheet" href="/css/custom.css" />
     <script src="/js/jquery.js"></script>
     <script src="/js/bootstrap.js"></script>
 </head>
@@ -16,7 +16,7 @@
         <a class="navbar-brand"><h2 class="logo">KakeiboApp</h2></a>
         <div class="account">
             <ul>
-                <li class="account-link"><img src="../static/img/user.ico" width="60%"></li>
+                <li class="account-link"><img src="/img/user.ico" width="60%"></li>
                 <li class="account-link"><span>ユーザー さん</span></li>
             </ul>
         </div>
@@ -28,84 +28,91 @@
         <!-- left side bar -->
         <div class="col sidebar">
             <ul class="nav flex-column nav-pills">
-                <li class="nav-item"><a class="nav-link" href="../templates/new.html">入力</a></li>
-                <li class="nav-item"><a class="nav-link" href="../templates/month.html">月別一覧</a></li>
-                <li class="nav-item"><a class="nav-link" href="../templates/year.html">年間一覧</a></li>
-                <li class="nav-item"><a class="nav-link active" href="../templates/setting.html">設定</a></li>
+                <li class="nav-item"><a class="nav-link" href="/new">入力</a></li>
+                <li class="nav-item"><a class="nav-link" href="/month">月別一覧</a></li>
+                <li class="nav-item"><a class="nav-link" href="/year">年間一覧</a></li>
+                <li class="nav-item"><a class="nav-link active" href="/setting">設定</a></li>
             </ul>
         </div>
 
         <!-- Main contents -->
-        <div class="col">
-            <div class="setting-header">
-                <h2>設定</h2>
-            </div>
+    <div class="col">
+        <div class="setting-header">
+            <h2>設定</h2>
+        </div>
 
-            <!-- Category Setting (outgo)-->
-            <div class="outgo-setting">
-                <p>カテゴリ(支出)</p>
+        <!-- Category Setting (outgo)-->
+        <div class="outgo-setting">
+            <p>カテゴリ(支出)</p>
+            <form th:action="@{/setting}" th:method="post">
                 <div class="form-group row">
                     <div class="col-sm-4">
-                        <input type="text" class="form-control">
+                        <input type="text" class="form-control" name="name">
                     </div>
                     <div class="col-sm-2">
                         <button class="btn btn-outline-secondary">追加</button>
                     </div>
                 </div>
+                <input type="hidden" name="type" value="outgo">
+            </form>
 
-                <table class="table">
-                    <thead>
-                    <tr>
-                        <th scope="col">カテゴリ</th>
-                        <th scope="col">色</th>
-                        <th scope="col"></th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    <tr>
-                        <td>食費</td>
-                        <td>橙</td>
-                        <td>
-                            <button class="btn btn-outline-secondary edit">編集</button>
-                            <button class="btn btn-outline-secondary delete">削除</button>
-                        </td>
-                    </tr>
-                    </tbody>
-                </table>
-            </div>
 
-            <!-- Category Setting (income)-->
-            <div class="outgo-setting">
-                <p>カテゴリ(収入)</p>
+            <table class="table">
+                <thead>
+                <tr>
+                    <th scope="col">カテゴリ</th>
+                    <th scope="col">色</th>
+                    <th scope="col"></th>
+                </tr>
+                </thead>
+                <tbody>
+                <tr th:each="category:${categories_outgo}" th:object="${category}">
+                    <td th:text="*{name}"></td>
+                    <td>橙</td>
+                    <td>
+                        <button class="btn btn-outline-secondary edit">編集</button>
+                        <button class="btn btn-outline-secondary delete">削除</button>
+                    </td>
+                </tr>
+                </tbody>
+            </table>
+        </div>
+
+        <!-- Category Setting (income)-->
+        <div class="outgo-setting">
+            <p>カテゴリ(収入)</p>
+            <form th:actoon="@{/setting}" method="post">
                 <div class="form-group row">
                     <div class="col-sm-4">
-                        <input type="text" class="form-control">
+                        <input type="text" class="form-control" name="name">
                     </div>
                     <div class="col-sm-2">
                         <button class="btn btn-outline-secondary">追加</button>
                     </div>
                 </div>
+                <input type="hidden" name="type" value="income">
+            </form>
 
-                <table class="table">
-                    <thead>
-                    <tr>
-                        <th scope="col">カテゴリ</th>
-                        <th scope="col">色</th>
-                        <th scope="col"></th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    <tr>
-                        <td>食費</td>
-                        <td>橙</td>
-                        <td>
-                            <button class="btn btn-outline-secondary edit">編集</button>
-                            <button class="btn btn-outline-secondary delete">削除</button>
-                        </td>
-                    </tr>
-                    </tbody>
-                </table>
-            </div>
+            <table class="table">
+                <thead>
+                <tr>
+                    <th scope="col">カテゴリ</th>
+                    <th scope="col">色</th>
+                    <th scope="col"></th>
+                </tr>
+                </thead>
+                <tbody>
+                <tr th:each="category:${categories_income}" th:object="${category}">
+                    <td th:text="*{name}"></td>
+                    <td>橙</td>
+                    <td>
+                        <button class="btn btn-outline-secondary edit">編集</button>
+                        <button class="btn btn-outline-secondary delete">削除</button>
+                    </td>
+                </tr>
+                </tbody>
+            </table>
+        </div>
 
 
         </div>
